Fix to pass norme v3

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tsierra- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/11 12:15:52 by tsierra-          #+#    #+#             */
-/*   Updated: 2021/03/02 15:49:55 by tsierra-         ###   ########.fr       */
+/*   Updated: 2021/03/04 17:24:31 by tsierra-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@
 #  define FD_SIZE 4096
 # endif
 
-typedef struct		s_list
+typedef struct s_list
 {
 	void			*content;
 	struct s_list	*next;
@@ -72,7 +72,7 @@ void		ft_putstr_fd(char *s, int fd);
 void		ft_putendl_fd(char *s, int fd);
 void		ft_putnbr_fd(int n, int fd);
 int			get_next_line(int fd, char **line);
-int			ft_nbrlen(long int nbr);
+size_t		ft_nbrlen(long int nbr);
 int			ft_strcmp(const char *s1, const char *s2);
 void		free_2d(char **ptr);
 int			ft_is_str_digit(char *str);
